% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{attachRkekuleDependencies}
\alias{attachRkekuleDependencies}
\title{Explicitly Include The HTML Dependencies}
\usage{
attachRkekuleDependencies(x = div())
}
\arguments{
\item{x}{object (tag) to attach the dependencies to. Default to an empty div}
}
\value{
HTML empty div tag with rkekule dependencies attached
}
\description{
Attaches the rkekule dependencies to a HTML tag or widget.
}
