% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemViewer.R
\name{chemViewer}
\alias{chemViewer}
\title{Chem Viewer Widget}
\usage{
chemViewer(data, width = NULL, height = NULL, chemicalFormat = c("mol",
  "cml"), renderType = c("2D", "3D"), showTools = TRUE)
}
\arguments{
\item{data}{moltable or cml}

\item{width}{widget width}

\item{height}{widget height}

\item{chemicalFormat}{one of \code{c("mol", "cml"}}

\item{renderType}{one of \code{c("2D", "3D")}}

\item{showTools}{show toolbar.}
}
\value{
htmlwidget
}
\description{
Create a widget to view chemical structure data in 2D or 3D.
Basic interactions like zooming and rotation are supported.
}
\details{
Limitations: The current version only supports moltables and cml.
}
