% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readCML}
\alias{readCML}
\title{Read CML structure data from file}
\usage{
readCML(file, nchars = NULL, encoding = "UTF-8")
}
\arguments{
\item{file}{CML file, either a character string (path) or a connection}

\item{nchars}{number of characters to read from the file or connection}

\item{encoding}{encoding of the CML file to read (default value \code{"UTF-8"})}
}
\value{
\code{character()} of length 1
}
\description{
Read CML structure data from file
}
