% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readMolfile}
\alias{readMolfile}
\title{Read Molfile structure data from file}
\usage{
readMolfile(file, nchars = NULL, encoding = "UTF-8")
}
\arguments{
\item{file}{Molfile, either a character string (path) or a connection}

\item{nchars}{number of characters to read from the file or connection}

\item{encoding}{encoding of the molfile to read (default value \code{"UTF-8"})}
}
\value{
\code{character()} of length 1
}
\description{
Read Molfile structure data from file
}
