% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderRkekule}
\alias{renderRkekule}
\title{Widget render function for use in Shiny}
\usage{
renderRkekule(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}
