% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkekule.R
\docType{package}
\name{rkekule}
\alias{rkekule}
\alias{rkekule-package}
\title{Wrapper of the JavaScript library \code{Kekule.js}}
\description{
Render interative visualizations of chemical structures using
the \code{Kekule.js} JavaScript library and the htmlwidgets package.
}
\details{
\enumerate{
	\item use \code{\link{chemViewer}} to create an html widget
 \item use \code{\link{renderRkekule}} and \code{\link{rkekuleOutput}} to
 use a widget in Shiny.
}
}
