% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{rkekuleColumnDefs}
\alias{rkekuleColumnDefs}
\title{Provides a html wrapper for a table column with chemical structure data.}
\usage{
rkekuleColumnDefs(targets, useTooltip = TRUE)
}
\arguments{
\item{targets}{vector of zero-based indices of columns that need to be rendered}

\item{useTooltip}{only show the structure upon hovering}
}
\description{
Return an expression to use in the \code{options$coldefs}
argument of \link[DT]{datatable}.
}
