% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{rkekuleFnDrawCallback}
\alias{rkekuleFnDrawCallback}
\title{Provides a javascript call to draw the html-wrapper structures}
\usage{
rkekuleFnDrawCallback(settings = "type: 'mol', width: 100, height: 100")
}
\arguments{
\item{settings}{kekule rendering settings}
}
\value{
HTML script tag
}
\description{
Returns an expression to use in the \code{options$fnDrawCallback}
argument of \link{DT::datatable}.
}
