% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlBinding.R
\name{rkekuleHtmlBinding}
\alias{rkekuleHtmlBinding}
\title{Create a Kekule.js HTML binding with inline data definition.}
\usage{
rkekuleHtmlBinding(dataChemObj, id, width = 500, height = 400)
}
\arguments{
\item{dataChemObj}{the data}

\item{id}{avoid clashes between multiple bindings}
}
\value{
\code{character()} vector of length 1
}
\description{
Generate a string of HTML code with the structure data included
inline that will be initialized by HTML binding. This is useful in an embedded
setting where the structure data is available in the context where the UI
is constructed. E.g. when inside renderUI in shiny.
}
\references{
https://partridgejiang.github.io/Kekule.js/documents/tutorial/content/chemViewer.html#creating-viewer
}
