% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rkekuleOutput}
\alias{rkekuleOutput}
\title{Widget Output Function For Use In Shiny}
\usage{
rkekuleOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget Output Function For Use In Shiny
}
