% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readCML}
\alias{readCML}
\title{Read CML structure data from file}
\usage{
readCML(fileName, con = stdin())
}
\arguments{
\item{fileName}{absolute or relative path}
}
\value{
\code{character()} of length 1
}
\description{
Read CML structure data from file
}
