% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readMolfile}
\alias{readMolfile}
\title{Read Molfile structure data from file}
\usage{
readMolfile(fileName, con = stdin())
}
\arguments{
\item{fileName}{absolute or relative path}
}
\value{
\code{character()} of length 1
}
\description{
Read Molfile structure data from file
}
