% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rkekuleOutput}
\alias{rkekuleOutput}
\alias{renderRkekule}
\title{Helper functions for using chemViewer in Shiny}
\usage{
rkekuleOutput(outputId, width = "100\%", height = "400px")

renderRkekule(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read the widget from}

\item{width}{the width of the widget container}

\item{height}{the height of the widget container}

\item{expr}{output of \code{\link{chemViewer}}}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with \code{quote()})?
This is useful if you want to save an expression in a variable.}
}
\description{
\code{rkekuleOutput}: create a container for the widget

\code{renderRkekule}: server logic to render the widget
}
